Generators
Generators are a powerful new kind of function that is able to pause its execution while also maintaining its own state. Generators are great for iterating over a list of items one at a time so you can handle each item on its own before moving on to the next one. You can also use generators to handle nested callbacks. For example, let's say that an app needs to get a list of all repositories and the number of times they've been starred. Well, before you can get the number of stars for each repository, you'd need to get the user's information. Then after retrieving the user's profile the code can then take that information to find all of the repositories.

More examples
https://codeburst.io/understanding-generators-in-es6-javascript-with-examples-6728834016d5
https://medium.com/dailyjs/a-simple-guide-to-understanding-javascript-es6-generators-d1c350551950